requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-RESTful==0.3.9
python-dotenv==0.19.2
SQLAlchemy==1.4.31
alembic==1.13.2
psycopg2==2.9.9
flask_jwt_extended